Initial Upload
This is the initial upload of my current website before things get
edited and updated. Need to move things around on my server first in
order for this to work well with my current image uploading software as
well as additions to my website in a easy to update way.